backend/requirements.txt
# Core dependencies for Turerz backend
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.0
pytz==2023.3

# MCP (Model Context Protocol) dependencies
mcp==1.0.0
httpx==0.27.0
anyio>=4.6

# OpenAI for natural language processing
openai==1.54.3

# Browser automation (migrating to Playwright)
playwright==1.48.0
selenium==4.25.0  # Keep for gradual migration

# CLI and utilities
typer==0.9.0
rich==13.7.0

# Data processing
pandas==2.3.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
